refactor(dashboard): split dashboard_data_loader; record cycle delta
dashboard_data_loader.py was 930 lines, over the 800-line coding-style
cap. Extract the self-contained bottom section (Prowler analysis +
provider finding filters + environment status) into
dashboard_data_analysis.py (220 lines) via façade re-export.

- dashboard_data_loader.py: now 735 lines; re-exports analyze_prowler,
  the eight *_findings provider filters, get_env_status, and
  _normalize_severity (the pure-helper test reaches it through the
  loader namespace) with `# noqa: F401`. Public API and __all__
  unchanged; the importer (dashboard-gen.py) and both loader test
  modules need no change.

Also records the #283–#284 cycle (DAST signal quality + dashboard_mapping
split) and this data_loader split in MEMORY.md, and documents why
diagram-gen.py (hyphenated CLI loaded by test via spec_from_file_location
with no sys.path; internal import-cycle risk) and checks.sh (core Bash,
kcov-gated) are deliberately NOT split here — both are only ~16% over a
soft cap and do not fit the Python façade pattern cleanly.

Verification:
- analysis body preserved byte-for-byte (diff-checked).
- pytest scanner/tests/ -> 1368 passed.
- scanner/lib coverage 99.12% (>= 99% floor); both modules covered.
- test_ci_*.py -> 195 passed; markdownlint clean on MEMORY.md + AGENTS.md.
